Rename: 청년정책, 청년공간 공통 ui 컴포넌트 폴더 이동
