+++ src/app/(hasMap)/youth-place/_components/ScrollablePlaces/ScollablePlaces.module.scss
