+++ src/app/(hasMap)/youth-place/_components/SideBar/index.tsx
@@ -1,13 +1,13 @@
 import styles from './SideBar.module.scss';
-import Filter from '@/app/(hasNotMap)/youth-place/_components/Filter';
-import ToggleButton from '@/app/(hasNotMap)/youth-place/_components/ToggleButton';
-import ScrollablePlaces from '@/app/(hasNotMap)/youth-place/_components/ScrollablePlaces';
+import PlaceFilter from '@/components/policy_place/PlaceFilter';
+import ToggleButton from '@/components/policy_place/ToggleButton';
+import ScrollablePlaces from '@/app/(hasMap)/youth-place/_components/ScrollablePlaces';
 
 export default function SideBar() {
   return (
     <section className={styles.container}>
       <ToggleButton defaultView="map" className={styles.toggleBtn} />
-      <Filter />
+      <PlaceFilter />
       <ScrollablePlaces />
     </section>
   );
