+++ src/app/(hasNotMap)/youth-place/(card)/page.tsx
@@ -1,6 +1,6 @@
 import styles from './page.module.scss';
-import Filter from '../_components/Filter';
-import ToggleButton from '../_components/ToggleButton';
+import PlaceFilter from '../../../../components/policy_place/PlaceFilter';
+import ToggleButton from '../../../../components/policy_place/ToggleButton';
 import FilterablePlaces from '../_components/FilterablePlaces';
 import Banner from '../_components/Banner';
 
@@ -8,7 +8,7 @@ export default function PlacePage() {
   return (
     <section className={styles.pageContainer}>
       <Banner />
-      <Filter />
+      <PlaceFilter />
       <ToggleButton defaultView="card" />
       <FilterablePlaces />
     </section>
