+++ src/app/(hasNotMap)/youth-place/[placeId]/page.tsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { useRef } from 'react';
-import PlaceMainInfo from '@/app/(hasNotMap)/youth-place/_components/PlaceMainInfo';
 import styles from './page.module.scss';
+import PlaceMainInfo from '@/app/(hasNotMap)/youth-place/_components/PlaceMainInfo';
 import Gallery from '@/app/(hasNotMap)/youth-place/_components/Gallery';
 import PlaceDetailInfo from '@/app/(hasNotMap)/youth-place/_components/PlaceDetailInfo';
 import WayToGo from '@/app/(hasNotMap)/youth-place/_components/WayToGo';
 import ButtonWrapper from '@/app/(hasNotMap)/youth-place/_components/ButtonWrapper';
-import ActionToolbars from '../../_components/ActionToolbars';
-import DetailContents from '../../_components/DetailContents';
+import ActionToolbars from '../../../../components/policy_place/ActionToolbars';
+import DetailContents from '../../../../components/policy_place/DetailContents';
 
 export default function PlaceDetailPage() {
   const componentRef = useRef<HTMLDivElement>(null);
