+++ src/components/policy_place/DetailContents/DetailContents.module.scss
