+++ src/components/policy_place/DetailContents/index.tsx
