+++ src/components/policy_place/IconCounter/IconCounter.module.scss
