+++ src/components/policy_place/IconCounter/index.tsx
